Change: type of Slice time to double
`double` 型でスライスの時刻を管理する.
後にスライスの時刻の間隔に応じて,
スライス同士の距離を変えることも検討する

+++ Slice.hpp
@@ -5,7 +5,7 @@
 #include <limits>
 
 struct Slice {
-    int t;
+    double t;
     struct Point2 {
         double x, y;
     };
